refactor: move runCode to app.js
